moved _check_realm_and_write_access function to vlob_queries/utils.py
